refactor(www): simplify school sidebar ownership
